requirements.txt
torch==2.6.0
torchvision==0.21.0
torchaudio==2.6.0
torchmetrics==0.11.4
pydicom==2.3.1
SimpleITK==2.5.3
nibabel==4.0.2
numpy==1.24.3
pandas==1.5.3
scikit-learn==1.2.2
scikit-image==0.20.0
opencv-python==4.7.0.72
Pillow==9.5.0
matplotlib==3.7.1
seaborn==0.12.2
plotly==5.14.0
tqdm==4.65.0
PyYAML==6.0
python-dotenv==1.0.0
jupyter==1.0.0
ipywidgets==8.0.6
tensorboard==2.13.0
gputil==1.4.0
pytest==7.3.1
pytest-cov==4.1.0
